made several changes to get it working on gpu. especially used set_inctensor instead as set_subtensor had problems on gpu. maybe this is a known bug. also replaced **2 by T.pow as it earlier had problems with derivates when running on GPU

+++ ac.py
@@ -1,5 +1,5 @@
 import numpy as np
-
+import pdb
 import theano
 import theano.tensor as T
 theano.config.exception_verbosity = 'high'
@@ -19,9 +19,8 @@ class AC(object):
         :param activation: Non linearity to be applied in the hidden
                            layer. If `None`, output will be just template.
         """
-
         if template is None:
-            template_values = np.asarray(rng.uniform(0, 1, (template_size, template_size)))
+            template_values = np.float32(np.asarray(rng.uniform(0, 1, (template_size, template_size))))
             template = theano.shared(value=template_values, name='template', borrow=True)
         self.template = template
